git-svn-id: https://r7/svn/WebGL@160 45b56d4a-f8b8-b242-b192-c14f84797ed6

+++ Shaders/FluidSim/3D/advect_mass.fs.glsl
@@ -52,7 +52,8 @@ varyin vec2 TexCoords;
 #include "fluidsim3d_include"
 // #define MassResolution
 
-float zToWorldSpace(int z){ return float(z) * (float(Resolution.z) / float(MassResolution.z)); }
+// float zToWorldSpace(int z){ return float(z) * (float(Resolution.z) / float(MassResolution.z)); }
+int zToWorldSpace(int z){ return int(float(z) * (float(Resolution.z) / float(MassResolution.z))); }
 
 out_dim advect(sampler3D u, sampler3D m, vec3 x, float dt){
 	vec4 v = samplePoint(u, x); //sample point
